refactor(table) Extract pagination component from Paged table
